Added start/ restart button
Created a method in gameController to reset the gameboard

The gameboard const variable in the gameController is a reference of the
array of the board object. Hence it should be noted that mutating the
gameboard variable will also change the array in the board object.

Splice arr method is used to directly mutate the gameboard since its
reference cannot be changed due to the nature of const variable.